requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
anthropic>=0.25.0
jinja2>=3.1.0
python-dotenv>=1.0.0
